Move pairwise impl and patch `testing.product_dict`
