+++ tests/chainer_tests/functions_tests/evaluation_tests/test_accuracy.py
@@ -1,5 +1,3 @@
-import itertools
-import typing as tp  # NOQA
 import unittest
 
 import numpy
@@ -12,52 +10,6 @@ from chainer import testing
 from chainer.testing import attr
 from chainer.utils import force_array
 from chainer.utils import type_check
-
-
-def _testing_pairwise(
-        *parameters  # type: tp.Iterable[tp.Dict[str, tp.Any]]
-):
-    # type: (...) -> tp.Iterator[tp.Dict[str, tp.Any]]
-    """Generate combinations that cover all pairs.
-
-    The argument is the same as `chainer.testing.product_dict`.
-
-    """
-    parameter_lists = [list(dicts) for dicts in parameters]  # type: tp.List[tp.List[tp.Dict[str, tp.Any]]]  # NOQA
-
-    for nd_index in sorted(_nd_indices_to_cover_each_2d(
-            [len(dicts) for dicts in parameter_lists])):
-        yield {
-            k: v
-            for i, dicts in zip(nd_index, parameter_lists)
-            for k, v in dicts[i].items()}
-
-
-def _nd_indices_to_cover_each_2d(shape):
-    # type: (tp.Sequence[int]) -> tp.Iterator[tp.Tuple[int, ...]]
-    rs = numpy.random.RandomState(seed=0)
-    n = len(shape)
-    indices = [list(range(length)) for length in shape]  # type: tp.List[tp.List[int]]  # NOQA
-
-    # `(k_i, k_j) in uncovered[(i, j)]` iff it has not been yielded
-    # `nd_index` such that `(nd_index[i], nd_inde[j]) == (k_i, k_j)`.
-    uncovered = {}  # type: tp.Dict[tp.Tuple[int, int], tp.Set[tp.Tuple[int, int]]]  # NOQA
-    for i, j in itertools.combinations(range(n), 2):
-        uncovered[(i, j)] = set(itertools.product(indices[i], indices[j]))
-
-    nd_indices = list(itertools.product(*indices))  # type: tp.List[tp.Tuple[int, ...]]  # NOQA
-    rs.shuffle(nd_indices)
-    for nd_index in nd_indices:
-        count = 0
-        for i, j in itertools.combinations(range(n), 2):
-            try:
-                uncovered[(i, j)].remove((nd_index[i], nd_index[j]))
-            except KeyError:
-                pass
-            else:
-                count += 1
-        if count > 0:
-            yield nd_index
 
 
 def accuracy(x, t, ignore_label):
@@ -86,7 +38,7 @@ def accuracy(x, t, ignore_label):
 
 
 @testing.parameterize(
-    *_testing_pairwise(
+    *testing.product_dict(
         [{'x_shape': (10, 3), 't_shape': (10,)},
          {'x_shape': (10, 3, 1), 't_shape': (10,)},
          {'x_shape': (10, 3, 1, 1), 't_shape': (10,)},
